Version 1
Currently using PRAW, the top 100 posts in r/EarthPorn (beautiful photos of nature subreddit) are put into a list, then iterated through randomly to select useable posts. The image is downloaded, opened, prints "FeelsGoodMan" in the center, saved, then uploaded to Twitter @FeelsGoodManBot. The program then sleeps for 1 hour, and attempts to repeat the same process until the list of 100 posts is empty. If the attempt fails, the program sleeps for 15 minutes (Twitter throttles post frequency).